pyproject.toml
[project]
name = 'av-backend'
version = '0.1.0'
description = 'Система контроля общественного транспорта'

dependencies = [
    'django-cors-headers',
    'django-environ',
    'gunicorn',
    'uvicorn',
    'pip-tools',
    'pre-commit',
    'pyproj',
    'django',
    'djangorestframework',
    'django-filter',
    'drf-spectacular',
    'psycopg2-binary',
]

[tool.setuptools]
py-modules = []

[tool.isort]
profile = 'black'
line_length = 120
lines_after_imports = 2
lines_between_sections = 1
lines_between_types = 0
force_alphabetical_sort_within_sections = true
ensure_newline_before_comments = true
combine_as_imports = true
combine_star = true
skip_gitignore = true
balanced_wrapping = true

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.flake8]
max-line-length = 120

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'av.pytest_settings'
addopts = '--new-first --failed-first --nomigrations -vv'
filterwarnings = [
    'ignore::RuntimeWarning',
    'ignore::DeprecationWarning',
]
